refactor(saves): decompose slots/service into sub-modules (#558)
services/saves/slots/service.py was 770 LOC over the [ours] 700
threshold. Splits into focused internal sub-modules within the
slots/ package. SlotsService class stays the public entry point;
sub-modules are internal helpers instantiated from the config.

Tests unchanged. Behavior preserved.

Closes #558
